apps/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
redis==5.0.1
rq==1.16.1
sse-starlette==2.0.0
python-multipart==0.0.9
httpx==0.27.0
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-httpx==0.30.0
factory-boy==3.3.0
faker==22.5.0
